requirements/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-environ==0.9.0
docopt==0.6.2
idna==3.3
packaging==21.3
Pillow==9.2.0
pip-review==1.1.1
pipdeptree==2.2.1
pipreqs==0.4.11
psycopg2-binary==2.9.3
pyparsing==3.0.9
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
yarg==0.1.9
